Added test endpoints as separate files. Fixed tests on Linux.
