+++ Sources/Endpoints/Endpoint+URLRequest.swift
@@ -1,11 +1,16 @@
 //
-//  File.swift
-//  
+//  Endpoint+URLRequest.swift
+//  Endpoints
 //
-//  Created by Zac White on 3/6/21.
+//  Created by Zac White on 1/26/19.
+//  Copyright © 2019 Velos Mobile LLC. All rights reserved.
 //
 
 import Foundation
+
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
 
 extension Endpoint {
