+++ Sources/Endpoints/Environment.swift
@@ -1,11 +1,16 @@
 //
-//  File.swift
-//  
+//  Environment.swift
+//  Endpoints
 //
-//  Created by Zac White on 3/6/21.
+//  Created by Zac White on 1/26/19.
+//  Copyright © 2019 Velos Mobile LLC. All rights reserved.
 //
 
 import Foundation
+
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
 
 public protocol EnvironmentType {
     /// The baseUrl of the Environment
