+++ Sources/Endpoints/Extensions/URLSession+Combine.swift
@@ -7,6 +7,8 @@
 //
 
 import Foundation
+
+#if canImport(Combine)
 import Combine
 
 @available(iOS 13.0, tvOS 13.0, watchOS 6.0, macOS 10.15, *)
@@ -115,3 +117,5 @@ extension URLSession {
             .eraseToAnyPublisher()
     }
 }
+
+#endif
