+++ Tests/EndpointsTests/Endpoints/Environment.swift
@@ -1,0 +1,16 @@
+//
+//  Environment.swift
+//  EndpointsTests
+//
+//  Created by Zac White on 1/26/19.
+//  Copyright © 2019 Velos Mobile LLC. All rights reserved.
+//
+
+import Foundation
+@testable import Endpoints
+
+struct Environment: EnvironmentType {
+    let baseUrl: URL
+
+    static let test = Environment(baseUrl: URL(string: "https://velosmobile.com")!)
+}
