+++ Tests/EndpointsTests/EndpointsTests.swift
@@ -6,162 +6,8 @@
 //  Copyright © 2019 Velos Mobile LLC. All rights reserved.
 //
 
-#if !os(watchOS)
 import XCTest
 @testable import Endpoints
-
-struct SimpleEndpoint: Endpoint {
-    static var definition: Definition<SimpleEndpoint> = Definition(
-        method: .get,
-        path: "user/\(path: \.name)/\(path: \.id)/profile"
-    )
-
-    struct Response: Decodable {
-        let response1: String
-    }
-
-    struct PathComponents {
-        let name: String
-        let id: String
-    }
-
-    let pathComponents: PathComponents
-}
-
-struct JSONProviderEndpoint: Endpoint {
-
-    static var definition: Definition<JSONProviderEndpoint> = Definition(
-        method: .get,
-        path: "user/\(path: \.name)/\(path: \.id)/profile"
-    )
-
-    struct Response: Decodable {
-        let responseOne: String
-    }
-
-    struct Body: Encodable {
-        let bodyValueOne: String
-    }
-
-    struct PathComponents {
-        let name: String
-        let id: String
-    }
-
-    let body: Body
-    let pathComponents: PathComponents
-
-    static let responseDecoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-
-    static let bodyEncoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.keyEncodingStrategy = .convertToSnakeCase
-        return encoder
-    }()
-}
-
-struct UserEndpoint: Endpoint {
-    static var definition: Definition<UserEndpoint> = Definition(
-        method: .get,
-        path: "hey" + \UserEndpoint.PathComponents.userId,
-        parameters: [
-            .form("string", path: \UserEndpoint.ParameterComponents.string),
-            .form("date", path: \UserEndpoint.ParameterComponents.date),
-            .form("double", path: \UserEndpoint.ParameterComponents.double),
-            .form("int", path: \UserEndpoint.ParameterComponents.int),
-            .form("bool_true", path: \UserEndpoint.ParameterComponents.boolTrue),
-            .form("bool_false", path: \UserEndpoint.ParameterComponents.boolFalse),
-            .form("time_zone", path: \UserEndpoint.ParameterComponents.timeZone),
-            .form("optional_string", path: \UserEndpoint.ParameterComponents.optionalString),
-            .form("optional_date", path: \UserEndpoint.ParameterComponents.optionalDate),
-            .formValue("hard_coded_form", value: "true"),
-            .query("string", path: \UserEndpoint.ParameterComponents.string),
-            .query("optional_string", path: \UserEndpoint.ParameterComponents.optionalString),
-            .query("optional_date", path: \UserEndpoint.ParameterComponents.optionalDate),
-            .queryValue("hard_coded_query", value: "true")
-        ],
-        headers: [
-            "HEADER_TYPE": .field(path: \UserEndpoint.HeaderComponents.headerValue),
-            "HARD_CODED_HEADER": .fieldValue(value: "test2"),
-            .keepAlive: .fieldValue(value: "timeout=5, max=1000")
-        ]
-    )
-
-    typealias Response = Void
-    
-    struct PathComponents {
-        let userId: String
-    }
-
-    struct ParameterComponents {
-        let string: String
-        let date: Date
-        let double: Double
-        let int: Int
-        let boolTrue: Bool
-        let boolFalse: Bool
-        let timeZone: TimeZone
-
-        let optionalString: String?
-        let optionalDate: String?
-    }
-
-    struct HeaderComponents {
-        let headerValue: String
-    }
-
-    let pathComponents: PathComponents
-    let parameterComponents: ParameterComponents
-    let headerComponents: HeaderComponents
-}
-
-struct PostEndpoint1: Endpoint {
-    static var definition: Definition<PostEndpoint1> = Definition(
-        method: .post,
-        path: "path"
-    )
-
-    typealias Response = Void
-
-    struct Body: Encodable {
-        let property1: Date
-        let property2: Int?
-    }
-
-    static let bodyEncoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.dateEncodingStrategy = .iso8601
-        return encoder
-    }()
-
-    let body: Body
-}
-
-struct PostEndpoint2: Endpoint {
-    static var definition: Definition<PostEndpoint2> = Definition(
-        method: .post,
-        path: "path"
-    )
-
-    typealias Response = Void
-
-    struct Body: Encodable {
-        let property1: String
-        let property2: Int?
-    }
-
-    let body: Body
-}
-
-struct Environment: EnvironmentType {
-    let baseUrl: URL
-
-    static let test = Environment(baseUrl: URL(string: "https://velosmobile.com")!)
-}
 
 class EndpointsTests: XCTestCase {
 
@@ -251,13 +97,4 @@ class EndpointsTests: XCTestCase {
             String(data: request.httpBody ?? Data(), encoding: .utf8)?.contains("double=2.3&int=42&bool_true=true&bool_false=false&time_zone=America/Los_Angeles&hard_coded_form=true") ?? false
         )
     }
-
-    static var allTests = [
-        ("testBasicEndpoint", testBasicEndpoint),
-        ("testBasicEndpointWithCustomDecoder", testBasicEndpointWithCustomDecoder),
-        ("testPostEndpointWithEncoder", testPostEndpointWithEncoder),
-        ("testPostEndpoint", testPostEndpoint),
-        ("testParameterEndpoint", testParameterEndpoint)
-    ]
 }
-#endif
