--- Tests/LinuxMain.swift
@@ -1,9 +1,0 @@
-#if !os(watchOS)
-import XCTest
-
-import EndpointsTests
-
-var tests = [XCTestCaseEntry]()
-tests += EndpointsTests.allTests()
-XCTMain(tests)
-#endif
